+++ src/Integration.php
@@ -7,8 +7,8 @@
  * Company: Pronamic
  *
  * @author Remco Tolsma
- * @version 1.1.0
- * @since 1.1.0
+ * @version 1.0.4
+ * @since 1.0.0
  */
 class Pronamic_WP_Pay_Gateways_Mollie_IDeal_Integration extends Pronamic_WP_Pay_Gateways_AbstractIntegration {
 	public function __construct() {
@@ -23,15 +23,22 @@ class Pronamic_WP_Pay_Gateways_Mollie_IDeal_Integration extends Pronamic_WP_Pay_
 		return 'Pronamic_WP_Pay_Gateways_Mollie_IDeal_ConfigFactory';
 	}
 
-	public function get_config_class() {
-		return 'Pronamic_WP_Pay_Gateways_Mollie_IDeal_Config';
-	}
-
 	public function get_settings_class() {
 		return 'Pronamic_WP_Pay_Gateways_Mollie_IDeal_Settings';
 	}
 
-	public function get_gateway_class() {
-		return 'Pronamic_WP_Pay_Gateways_Mollie_IDeal_Gateway';
+	/**
+	 * Get required settings for this integration.
+	 *
+	 * @see https://github.com/wp-premium/gravityforms/blob/1.9.16/includes/fields/class-gf-field-multiselect.php#L21-L42
+	 * @since 1.0.4
+	 * @return array
+	 */
+	public function get_settings() {
+		$settings = parent::get_settings();
+
+		$settings[] = 'mollie_ideal';
+
+		return $settings;
 	}
 }